Start to add actions and reducers.
