+++ src/services/actions/index.js
@@ -1,1 +1,10 @@
-export const A_B = 'A_B';
+export const SET_CURRENT_INGREDIENT_DETAILS_OBJECT = 'SET_CURRENT_INGREDIENT_DETAILS_OBJECT';
+
+export function setCurrentIngredientDetailsObject(id) {
+    return function (dispatch) {
+        dispatch({
+            type: SET_CURRENT_INGREDIENT_DETAILS_OBJECT,
+            id
+        });
+    };
+}
