+++ src/services/reducers/ingredientsOrder.js
@@ -1,5 +1,5 @@
 
-import { A_B } from '../actions';
+import { SET_CURRENT_INGREDIENT_DETAILS_OBJECT } from '../actions';
 
 const initialState = {
     dataIngredientsList: [],
@@ -10,6 +10,12 @@ const initialState = {
 
 export const ingredientsOrderReducer = (state = initialState, action) => {
     switch (action.type) {
+        case SET_CURRENT_INGREDIENT_DETAILS_OBJECT: {
+            return {
+                ...state, 
+                currentIngredientDetailsObject: [...state.dataIngredientsList].filter(item => item.id !== action.id)
+            };
+        }
         default: {
             return state;
         }
